login api

+++ src/main/java/com/nju/scrum/pojo/User.java
@@ -9,7 +9,7 @@ public class User {
     private String nickname;
     private String city;
     private Integer credit;//表示没有参与活动的次数 初始为0
-    private Integer unlockTime;//解封时间 初始化为0 即时间戳的初始时间1970年1月1日
+    private Long unlockTime;//解封时间 初始化为0 即时间戳的初始时间1970年1月1日
     public Integer getUid() {
         return uid;
     }
@@ -67,11 +67,11 @@ public class User {
         this.credit = credit;
     }
 
-    public Integer getUnlockTime() {
+    public Long getUnlockTime() {
         return unlockTime;
     }
 
-    public void setUnlockTime(Integer unlockTime) {
+    public void setUnlockTime(Long unlockTime) {
         this.unlockTime = unlockTime;
     }
